Move PlayerBeforeSpawnEvent and PlayerSpawnCompleteEvent to Shared
